+++ clients/admin-ui/src/features/integrations/integration-type-info/rdsMySQLInfo.tsx
@@ -21,6 +21,7 @@ export const RDS_MYSQL_TAGS = [
   "Database",
   "Detection",
   "Discovery",
+  "DSR automation",
   "RDS",
   "MySQL",
 ];